fix lint error and remove Redux/Actions #12
apply suggested change #22
