+++ src/Redux/index.js
@@ -1,5 +1,5 @@
 import { combineReducers } from 'redux';
-import authReducer from './authReducer';
-import postReducer from './postReducer';
+import authReducer from './auth';
+import postReducer from './post';
 
 export default combineReducers({ authReducer, postReducer });
