+++ src/Redux/post.js
@@ -1,3 +1,5 @@
+import { createAction, handleActions as postReducer } from 'redux-actions';
+
 import {
   INIT,
   POST_LIKE,
@@ -5,24 +7,104 @@ import {
   POST_ONE,
   POST_API,
 } from 'Constants/api-uri';
-import {
-  getAllPostsFail,
-  getAllPostsSuccess,
-  getUserPostSuccess,
-  getUserPostFail,
-  clickPostSuccess,
-  clickPostFail,
-  uploadPostSuccess,
-  uploadPostFail,
-  likePostSuccess,
-  likePostFail,
-  delikePostFail,
-  delikePostSuccess,
-} from '../Reducers/postReducer';
+
+const initialState = {
+  feed: null,
+};
+
+const GET_ALLPOSTS_SUCCESS = 'post/GET_ALLPOST_SUCCESS';
+const GET_ALLPOSTS_FAIL = 'post/GET_ALLPOST_FAIL';
+
+const GET_USER_POST_SUCCESS = 'post/GET_USER_POST_SUCCESS';
+const GET_USER_POST_FAIL = 'post/GET_USER_POST_SUCCESS';
+
+const CLICK_POST_SUCCESS = 'post/CLICK_POST_SUCCESS';
+const CLICK_POST_FAIL = 'post/CLICK_POST_FAIL';
+
+const UPLOAD_POST_SUCCESS = 'post/UPLOAD_POST_SUCCESS';
+const UPLOAD_POST_FAIL = 'post/UPLOAD_POST_FAIL';
+
+const DELETE_POST_SUCCESS = 'post/DELETE_POST_SUCCESS';
+const DELETE_POST_FAIL = 'post/DELETE_POST_FAIL';
+
+const LIKE_POST_SUCCESS = 'post/LIKE_POST_SUCCESS';
+const LIKE_POST_FAIL = 'post/LIKE_POST_FAIL';
+
+const DELIKE_POST_SUCCESS = 'post/DELIKE_POST_SUCCESS';
+const DELIKE_POST_FAIL = 'post/DELIKE_POST_FAIL';
+
+export const getAllPostsSuccess = createAction(GET_ALLPOSTS_SUCCESS); // result.posts
+export const getAllPostsFail = createAction(GET_ALLPOSTS_FAIL);
+export const getUserPostSuccess = createAction(GET_USER_POST_SUCCESS);
+export const getUserPostFail = createAction(GET_USER_POST_FAIL);
+export const clickPostSuccess = createAction(CLICK_POST_SUCCESS); // result.post
+export const clickPostFail = createAction(CLICK_POST_FAIL);
+export const uploadPostSuccess = createAction(UPLOAD_POST_SUCCESS); // result
+export const uploadPostFail = createAction(UPLOAD_POST_FAIL);
+export const deletePostSuccess = createAction(DELETE_POST_SUCCESS);
+export const deletePostFail = createAction(DELETE_POST_FAIL);
+export const likePostSuccess = createAction(LIKE_POST_SUCCESS); // result
+export const likePostFail = createAction(LIKE_POST_FAIL);
+export const delikePostSuccess = createAction(DELIKE_POST_SUCCESS); // result
+export const delikePostFail = createAction(DELIKE_POST_FAIL);
+
+export default postReducer(
+  {
+    [GET_ALLPOSTS_SUCCESS]: (state, action) => {
+      return { ...state, feed: action.payload };
+    },
+    [GET_ALLPOSTS_FAIL]: (state) => {
+      return state;
+    },
+
+    [GET_USER_POST_SUCCESS]: (state) => {
+      return state;
+    },
+    [GET_USER_POST_FAIL]: (state) => {
+      return state;
+    },
+
+    [CLICK_POST_SUCCESS]: (state) => {
+      return state;
+    },
+    [CLICK_POST_FAIL]: (state) => {
+      return state;
+    },
+
+    [UPLOAD_POST_SUCCESS]: (state) => {
+      return state;
+    },
+    [UPLOAD_POST_FAIL]: (state) => {
+      return state;
+    },
+
+    [DELETE_POST_SUCCESS]: (state) => {
+      return state;
+    },
+    [DELETE_POST_FAIL]: (state) => {
+      return state;
+    },
+
+    [LIKE_POST_SUCCESS]: (state) => {
+      return state;
+    },
+    [LIKE_POST_FAIL]: (state) => {
+      return state;
+    },
+
+    [DELIKE_POST_SUCCESS]: (state) => {
+      return state;
+    },
+    [DELIKE_POST_FAIL]: (state) => {
+      return state;
+    },
+  },
+  initialState,
+);
 
 export const getAllPosts = (token) => {
-  // eslint-disable-next-line consistent-return
   return async (dispatch) => {
+    let result = null;
     try {
       const response = await fetch(
         `${process.env.REACT_APP_SERVER_URL}${INIT}${POST_API}`,
@@ -33,23 +115,22 @@ export const getAllPosts = (token) => {
           },
         },
       );
-      let result = null;
       if (response.status === 200) {
         result = await response.json();
         await dispatch(getAllPostsSuccess(result.posts));
       } else {
         await dispatch(getAllPostsFail());
       }
-      return result ? result.posts : null;
     } catch (error) {
       dispatch(getAllPostsFail());
     }
+    return result.posts;
   };
 };
 
 export const getPostByUserId = (userId, token) => {
-  // eslint-disable-next-line consistent-return
   return async (dispatch) => {
+    let result = null;
     try {
       const response = await fetch(
         `${process.env.REACT_APP_SERVER_URL}${INIT}${POST_API}/${userId}`,
@@ -60,22 +141,22 @@ export const getPostByUserId = (userId, token) => {
           },
         },
       );
-      let result = null;
       if (response.status === 200) {
         result = await response.json();
         await dispatch(getUserPostSuccess());
       } else {
         await dispatch(getUserPostFail());
       }
-      return result ? result.posts : null;
     } catch (error) {
       dispatch(getUserPostFail());
     }
+    return result.posts;
   };
 };
 
 export const getPostByPostId = (postId, token) => {
   return async (dispatch) => {
+    let result = null;
     try {
       const response = await fetch(
         `${process.env.REACT_APP_SERVER_URL}${INIT}${POST_API}${POST_ONE}/${postId}`,
@@ -86,22 +167,22 @@ export const getPostByPostId = (postId, token) => {
           },
         },
       );
-      let result = null;
       if (response.status === 200) {
         result = await response.json();
         await dispatch(clickPostSuccess(result.post));
       } else {
         await dispatch(clickPostFail());
       }
-      return result ? result.post : null;
     } catch (error) {
       dispatch(clickPostFail());
     }
+    return result.post;
   };
 };
 
 export const uploadPost = (payload, token) => {
   return async (dispatch) => {
+    let result = null;
     try {
       const response = await fetch(
         `${process.env.REACT_APP_SERVER_URL}${INIT}${POST_API}`,
@@ -113,7 +194,6 @@ export const uploadPost = (payload, token) => {
           body: payload,
         },
       );
-      let result = null;
       if (response.status === 200) {
         result = await response.json();
         await dispatch(uploadPostSuccess(result));
@@ -127,6 +207,7 @@ export const uploadPost = (payload, token) => {
 
 export const uploadSketcherPost = (payload, token) => {
   return async (dispatch) => {
+    let result = null;
     try {
       const response = await fetch(
         `${process.env.REACT_APP_SERVER_URL}${INIT}${POST_API}${SKETCHER}`,
@@ -136,15 +217,14 @@ export const uploadSketcherPost = (payload, token) => {
           body: payload,
         },
       );
-      let result = null;
       if (response.status === 200) {
         result = await response.json();
         await dispatch(uploadPostSuccess(result));
       }
-      return { type: 'UPLOAD_POST_SUCCESS' };
     } catch (err) {
       dispatch(uploadPostFail());
     }
+    return { type: 'UPLOAD_POST_SUCCESS' };
   };
 };
 
@@ -170,6 +250,7 @@ export const deletePost = (id, token) => {
 
 export const postLike = (postId, token) => {
   return async (dispatch) => {
+    let result = null;
     try {
       const response = await fetch(
         `${process.env.REACT_APP_SERVER_URL}${INIT}${POST_API}${POST_LIKE}/${postId}`,
@@ -180,22 +261,22 @@ export const postLike = (postId, token) => {
           },
         },
       );
-      let result = null;
       if (response.status === 200) {
         result = await response.json();
         await dispatch(likePostSuccess(result));
       } else {
         await dispatch(likePostFail());
       }
-      return result;
     } catch (error) {
       dispatch(likePostFail());
     }
+    return result;
   };
 };
 
 export const deleteLike = (postId, token) => {
   return async (dispatch) => {
+    let result = null;
     try {
       const response = await fetch(
         `${process.env.REACT_APP_SERVER_URL}${INIT}${POST_API}${POST_LIKE}/${postId}`,
@@ -206,16 +287,15 @@ export const deleteLike = (postId, token) => {
           },
         },
       );
-      let result = null;
       if (response.status === 200) {
         result = await response.json();
         await dispatch(delikePostSuccess(result));
       } else {
         await dispatch(delikePostFail());
       }
-      return result;
     } catch (error) {
       dispatch(delikePostFail());
     }
+    return result;
   };
 };
